+++ app/Providers/AppServiceProvider.php
@@ -7,8 +7,10 @@ use App\Http\Responses\RegistrationResponse;
 use App\Listeners\HandleInvitationAcceptance;
 use App\Listeners\QueueCrystalUpdateListener;
 use App\Models\Invitation;
+use App\Models\Post;
 use App\Models\User;
 use App\Observers\InvitationObserver;
+use App\Observers\PostObserver;
 use App\Observers\UserObserver;
 use App\Policies\ContentPolicy;
 use App\Policies\InvitationPolicy;
@@ -61,6 +63,7 @@ class AppServiceProvider extends ServiceProvider
         // Register observers
         User::observe(UserObserver::class);
         Invitation::observe(InvitationObserver::class);
+        Post::observe(PostObserver::class);
 
         // Register Crystal update event listeners
         Event::listen(ContentPublishedEvent::class, [QueueCrystalUpdateListener::class, 'handleContentPublished']);